Move this code to its own file
